:truck: feat: rename various services
