+++ config/phpstan/silverstripe53.neon
@@ -1,4 +1,4 @@
 services:
     extensionOwnerMetaPropertyTypeResolver:
-        class: Cambis\SilverstripeRector\TypeResolver\TypeResolver\ExtensionOwnerMetaPropertyTypeResolver
+        class: Cambis\SilverstripeRector\TypeResolver\ExtensionOwnerMetaPropertyTypeResolver
         tags: [silverstan.typeResolver.propertyTypeResolver]
