+++ config/rector-services.php
@@ -5,14 +5,14 @@ declare(strict_types=1);
 use Cambis\Silverstan\Autoloader\Autoloader;
 use Cambis\Silverstan\ClassManifest\ClassManifest;
 use Cambis\Silverstan\TypeResolver\TypeResolver;
+use Cambis\SilverstripeRector\AnnotationComparator\AnnotationComparator;
+use Cambis\SilverstripeRector\AnnotationComparator\AnnotationComparator\ExtendsAnnotationComparator;
+use Cambis\SilverstripeRector\AnnotationComparator\AnnotationComparator\MethodAnnotationComparator;
+use Cambis\SilverstripeRector\AnnotationComparator\AnnotationComparator\MixinAnnotationComparator;
+use Cambis\SilverstripeRector\AnnotationComparator\AnnotationComparator\PropertyAnnotationComparator;
+use Cambis\SilverstripeRector\AnnotationComparator\AnnotationComparator\TemplateAnnotationComparator;
+use Cambis\SilverstripeRector\AnnotationComparator\Contract\AnnotationComparatorInterface;
 use Cambis\SilverstripeRector\Configuration\SilverstripeOption;
-use Cambis\SilverstripeRector\PhpDoc\AnnotationComparator;
-use Cambis\SilverstripeRector\PhpDoc\AnnotationComparator\ExtendsAnnotationComparator;
-use Cambis\SilverstripeRector\PhpDoc\AnnotationComparator\MethodAnnotationComparator;
-use Cambis\SilverstripeRector\PhpDoc\AnnotationComparator\MixinAnnotationComparator;
-use Cambis\SilverstripeRector\PhpDoc\AnnotationComparator\PropertyAnnotationComparator;
-use Cambis\SilverstripeRector\PhpDoc\AnnotationComparator\TemplateAnnotationComparator;
-use Cambis\SilverstripeRector\PhpDoc\Contract\AnnotationComparatorInterface;
 use Rector\Config\RectorConfig;
 use Rector\NodeTypeResolver\DependencyInjection\PHPStanServicesFactory;
