+++ rules/Silverstripe413/Rector/Class_/AddGetOwnerMethodAnnotationToExtensionRector.php
@@ -8,7 +8,7 @@ use Override;
 use PhpParser\Node\Stmt\Class_;
 use PHPStan\PhpDocParser\Ast\PhpDoc\PhpDocTagValueNode;
 use PHPStan\Type\Generic\GenericObjectType;
-use Symplify\RuleDocGenerator\ValueObject\CodeSample\ConfiguredCodeSample;
+use Symplify\RuleDocGenerator\ValueObject\CodeSample\CodeSample;
 use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
 use function array_key_exists;
 
@@ -20,7 +20,7 @@ final class AddGetOwnerMethodAnnotationToExtensionRector extends AbstractAddAnno
     #[Override]
     public function getRuleDefinition(): RuleDefinition
     {
-        return new RuleDefinition('Add missing dynamic annotations.', [new ConfiguredCodeSample(
+        return new RuleDefinition('Add missing dynamic annotations.', [new CodeSample(
             <<<'CODE_SAMPLE'
 class FooExtension extends \SilverStripe\Core\Extension
 {
@@ -35,10 +35,6 @@ class FooExtension extends \SilverStripe\Core\Extension
 {
 }
 CODE_SAMPLE
-            ,
-            [
-                self::SET_TYPE_STYLE => self::SET_INTERSECTION,
-            ]
         )]);
     }
