+++ src/AnnotationComparator/AnnotationComparator/MethodAnnotationComparator.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Cambis\SilverstripeRector\PhpDoc\AnnotationComparator;
+namespace Cambis\SilverstripeRector\AnnotationComparator\AnnotationComparator;
 
+use Cambis\SilverstripeRector\AnnotationComparator\Contract\AnnotationComparatorInterface;
 use Cambis\SilverstripeRector\NodeAnalyser\ClassAnalyser;
-use Cambis\SilverstripeRector\PhpDoc\Contract\AnnotationComparatorInterface;
 use Cambis\SilverstripeRector\ValueObject\SilverstripeConstants;
 use Override;
 use PhpParser\Node;
